Move `ReactTextInlineImageShadowNode` to the `text.internal` package
